% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{aou_species}
\alias{aou_species}
\title{American Ornithological Society official checklist of North American and Middle American Birds}
\format{
A data frame with 2186 rows and 3 variables:
\describe{
\item{species}{Latin name}
\item{common_name}{English common name}
\item{french_name}{French common name}
}
}
\source{
Chesser, R. T., S. M. Billerman, K. J. Burns, C. Cicero, J. L. Dunn, B. E. Hernández-Baños, A. W. Kratter, I. J. Lovette, N. A. Mason, P. C. Rasmussen, J. V. Remsen, Jr., D. F. Stotz, and K. Winker. 2021. Check-list of North American Birds (online). American Ornithological Society. http://checklist.aou.org/taxa
}
\usage{
aou_species
}
\description{
American Ornithological Society official checklist of North American and Middle American Birds
}
\keyword{datasets}
