% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{change_yaml_matter}
\alias{change_yaml_matter}
\title{Change YAML header content}
\usage{
change_yaml_matter(input_file, ..., output_file)
}
\arguments{
\item{input_file}{Input .Rmd file with YAML to be updated}

\item{output_file}{Output .Rmd filename with updated YAML (optional; if left blank, updated .Rmd file will be printed in the console.)}
}
\value{
A .Rmd file with updated YAML content, either saved to the \code{output_file} path location or printed in the console.
}
\description{
This function updates the YAML content of a given Rmd file.
YAML content is passed as a list via the dots argument of the function.
}
\references{
Code from: https://stackoverflow.com/a/66908611/1454785
}
