% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{cpf_report_params}
\alias{cpf_report_params}
\title{Report analysis parameters for central place foragers}
\format{
A data frame:
\describe{
\item{mb_project_num}{Numeric. Movebank project number.}
\item{report_title}{Text. Title used in PDF report.}
\item{file_name}{Text. File name used in all outputs.}
\item{spp}{Text. Species common name.}
\item{breeding_pairs}{Numeric. Number of breeding pairs at colony. Used in calculating number of birds expected to occur in high-use areas.}
\item{count_year}{Numeric. Year when most recent colony count was conducted.}
\item{minDist}{Numeric. Distance, in kilometers, a bird must travel from colony to be considered in away from colony. Passed to OPPtools::opp_get_trips in analysis.}
\item{maxDist}{Numeric. Distance, in kilometers, a bird must return to colony to be considered in a complete trip. Passed to OPPtools::opp_get_trips in analysis.}
\item{minDur}{Numeric. Time, in hours, a bird must be away from the colony to be considered in a complete trip. Passed to OPPtools::opp_get_trips in analysis.}
\item{gapTime}{Numeric. Time, in hours, between consecutive locations to be considered a gap in tracking. Passed to OPPtools::opp_get_trips in analysis.}
\item{gapDist}{Numeric. Distance, in kilometers, between consecutive locations to be considered a gap in tracking. Passed to OPPtools::opp_get_trips in analysis.}
\item{interpolateGaps}{Logical. Should gaps in location data be interpolated? Passed to OPPtools::opp_get_trips in analysis.}
\item{timestep}{Timestep used for interpolating GPS locations, as text (e.g. 5 min, 1 hour). Passed to OPPtools::opp_get_trips in analysis.}
\item{gridRes}{Numeric. Grid resultion, in kilometers, used in calculating utilizations disributions.}
\item{gridExtend}{Numeric. Distance, in kilometers, used to extend grid used in calculating utilizations disributions.}
\item{kernelSmoother}{Text. Type of kernel smoother used in analysis to calculate utilization distributions. Acceptable values are: href, href/2, and step}
\item{stage}{Text. Describing the breeding stage data apply to. Used in text of report.}
\item{dates}{Text. Describing the date range of the breeding stage for this population. Used in text of report.}
}
}
\usage{
cpf_report_params
}
\description{
This dataset includes parameters used in analysis to identify high-use areas for breeding colonial seabirds. Parameters
in this dataframe are passed to OPPtools::opp_render_diagnostic and OPPtools::opp_render_report to run analysis to identify trips away from the colony,
calculate individual utilization distributions, quantify overlap among individual distributions, and map high-use areas around the colony.
}
\keyword{datasets}
