% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{getDist}
\alias{getDist}
\title{Calculate the distance between consecutive points}
\usage{
getDist(lon, lat, lonlat = TRUE)
}
\arguments{
\item{lon}{Vector of longitudes}

\item{lat}{Vector of latitudes}

\item{lonlat}{If TRUE, coordinates should be in degrees; else they should represent planar ('Euclidean') space (e.g. units of meters)}
}
\value{
A vector of distances in meters.
}
\description{
Wrapper for raster::pointDistance that only requires input of a
vector of longitudes and a vector of latitudes. Default calculation assumes
data are in decimal degrees. If not, then set lonlat = FALSE. Compatible with
tidyverse.
}
