% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_bbmm}
\alias{opp_bbmm}
\title{Calculate Brownian Bridge Movement Model (bbmm) kernels}
\usage{
opp_bbmm(
  data,
  sig2 = 20,
  bridgeSmooth = 1,
  res,
  extendGrid = 10,
  useGappy = TRUE,
  interpolated = TRUE,
  showLiker = FALSE,
  ...
)
}
\arguments{
\item{data}{Trips, as output by either opp_get_trips or ctcrw_interpolation.}

\item{sig2}{Numeric. \code{sig2} parameter in the bbmm model, typically expressed as GPS error in meters.}

\item{bridgeSmooth}{Numeric. Parameter to over-smooth bbmm kernels.}

\item{res}{Numeric. Resolution in km of base kernel grid.}

\item{extendGrid}{Numeric. Distance (km) to expand grid beyond the bounding box of tracking data. Default 10km.}

\item{useGappy}{Logical (T/F). Should "Gappy" trips as identified by opp_get_trips be included in bbmm? Default TRUE. If FALSE, only trips identified as "Complete" are used. Note this parameter is ignore if using interpolated data, where by definition the gaps have been interpolated over.}

\item{interpolated}{Logical (T/F). If an output from ctcrw_interpolation is provided, should the interpolated data be used? Default is TRUE.}

\item{showLiker}{Logical (T/F). Show plot outputs from adehabitatHR::liker (used to calculate bbmm \code{sig1} parameter)? Default FALSE.}
}
\description{
This function is a wrapper of adehabitatHR::kernelbb
and calculates Brownian Bridge Movement Model (bbmm) kernels for
a given set of tracks.
}
