% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_find_minDist}
\alias{opp_find_minDist}
\title{Explore how different values of minDist influence total number of trips}
\usage{
opp_find_minDist(
  data,
  minDist_range = 1:10,
  returnBuff = 20,
  duration = 1,
  gapLimit = 100,
  gapTime = 1,
  gapDist = 10
)
}
\arguments{
\item{data}{Polygon output from track2KBA::findSite()}

\item{minDist_range}{Vector of ditances (km) from the colony to calculate how
minimum distance influences total number of long foraging trips.
Used to label trips as 'Non-trip'. Defaults to 5}

\item{returnBuff}{Outer distance (km) to capture trips that start and end
away from the colony. Used to label trips as 'Incomplete' and for assessing the
effect of minDist on total number of trips. Defaults to 20.}

\item{duration}{Minimum trip duration (hrs)}

\item{gapLimit}{Maximum time between points to be considered too large to be
a contiguous tracking event. Can be used to ensure that deployments on the
same animal in different years do not get combined into extra long trips.
Defaults to 100 days.}

\item{gapTime}{Time (hrs) between successive locations at which trips will be flagged as 'Gappy'.
Used in connection with gapDist, such that locations must be farther apart in
both time and space to be considered a gap.}

\item{gapDist}{Distance (km) between successive locations at which trips will be flagged as 'Gappy'.
Used in connection with gapTime, such that locations must be farther apart in
both time and space to be considered a gap.}
}
\description{
Calculates the number of trips that go farther than returnBuff for a range of
minimum distance values that a bird must travel from the colony to be considered in a trip.
This is useful for ensuring that the minimum trip distance separates all long trips.
}
\details{
With lower sampling frequency or if GPS fixes are not taken at or near the colony,
multiple foraging trips may be combined if minDist is too small. Look for the
minDist value where the number of trips begins to asymptote. Selected values should
be confirmed by calling opp_get_trips directly with showPlots = TRUE.
}
