% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_href}
\alias{opp_href}
\title{Calculate median href for kernel density estimates}
\usage{
opp_href(data)
}
\arguments{
\item{data}{SpatialPointsDataFrame containing projected tracking data,
with an ID field indicating unique trips}
}
\description{
Calculate median href for kernel density estimates
}
