% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_kernel}
\alias{opp_kernel}
\title{Calculate traditional kernel density estimates}
\usage{
opp_kernel(
  data,
  interpolated = TRUE,
  extendGrid = 10,
  smoother = "href",
  res = 1
)
}
\arguments{
\item{data}{Tracks to calculate kernels on. Accepts output from either opp_get_trips or ctcrw_interpolation.}

\item{interpolated}{Logical (T/F). If provided an output from ctcrw_interpolation, should the interpolated tracks
be used for kernel calculation? Default TRUE. This parameter is ignored if the function is provided data from opp_get_tracks.}

\item{extendGrid}{Numeric. Distance (km) to expand grid beyond the bounding box of tracking data. Default 10km.}

\item{smoother}{Smoother value used in kernel calculations. By default uses the calculated href value of the tracks.}

\item{res}{Grid resolution in sq km to use for kernel calculations.}
}
\description{
This function is a wrapper of track2KBA::estSpaceUse
(which itself is a wrapper of adehabitatHR::kernelUD) to
calculate a traditional kernel density estimates
on a given set of trips. The function accepts outputs
from either opp_get_trips or ctcrw_interpolation.
If provided an output from ctcrw_interpolation, the function
will calculate kernels using the interpolated tracks by default.
This behavior can be changed by setting \code{interpolated = FALSE}.
The default kernel smoother is calculated using the href method.
The default UD level is 50\%.
}
