% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_map}
\alias{opp_map}
\title{Plot raw tracks from Movebank download}
\usage{
opp_map(data)
}
\arguments{
\item{data}{Movebank data as returned by \code{opp_download_data()} or sites output from \code{opp_sites()}.}
}
\description{
Quickly plot Movebank data downloaded
using \code{opp_download_data} or \code{opp_sites} to visualize tracks or sites.
}
\examples{
 \dontrun{
data(murres)
opp_map(murres)
}
}
