% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_map_tracks}
\alias{opp_map_tracks}
\title{Creates maps of biologger tracks from output of opp2KBA, opp_get_trips, or ctcrw_interpolation}
\usage{
opp_map_tracks(
  tracks = my_interp$data,
  center = my_track2kba$site,
  zoom = NULL,
  coast_scale = 10,
  viridis_option = "D",
  show_locs = T
)
}
\arguments{
\item{tracks}{data.frame, sf points, or SpatialPointsDataFrame containing fields named ID, Longitude, and Latitude}

\item{center}{Data frame containing columns 'Longitude' and 'Latitude' in decimal degrees,
for plotting the colony or nest locations.}

\item{zoom}{NULL or numeric value from 1:16, indicating the zoom level for map. If left as NULL (default)
map extent will be defined by the bounding box of tracks. If numeric value is provided, bounding box will
be cenetered on the mean location in tracks at the zoom level provided.}

\item{coast_scale}{Mapping resolution for the coastline basemap. Must be one of: 10 - high resolution,
50 - medium resolution, 110 - low resolution.}

\item{viridis_option}{A character string indicating the colormap option to
use. Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}

\item{show_locs}{Logical. Should biologger locations be plotted as points?}
}
\value{
A ggplot object
}
\description{
Creates maps of biologger tracks from output of opp2KBA, opp_get_trips, or ctcrw_interpolation
}
