% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_movebank_key}
\alias{opp_movebank_key}
\title{Set Movebank credentials}
\usage{
opp_movebank_key(username, set_keyring = FALSE)
}
\arguments{
\item{username}{Your Movebank username.}

\item{set_keyring}{Logical (T/F). Set to TRUE if you would like to further password protect your Movebank credentials. You will be prompted for this password immediately upon running the function.}
}
\description{
This function uses the R library \code{keyring} to interactively
and securely set your Movebank credentials for other functions to use.
The credentials are encrypted and stored under the name "OPP-Movebank" in
Keychain (Mac) or Credential Store (Windows). If you store your credentials
in this way, other OPP functions will no longer prompt you for your Movebank
username and password each time they are run.

For added security, you can further set a keychain password. This means
you will need to enter your keychain password any time an \code{OPPtools} function
attempts to access your stored Movebank credentials. By default this is turned
off, as the credentials are securely encrypted by \code{keyring}.
}
