% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPPlots.R
\name{opp_plot_repAssess}
\alias{opp_plot_repAssess}
\title{Custom plot of representativeness assessment from track2KBA::repAssess}
\usage{
opp_plot_repAssess(represent, plot = TRUE)
}
\value{
A ggplot object showing the results of the call to repAssess
}
\description{
Custom plot of representativeness assessment from track2KBA::repAssess
}
