% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPReporting.R
\name{opp_render_report}
\alias{opp_render_report}
\title{Render KBA report for a given species}
\usage{
opp_render_report(
  params,
  iterations = 5,
  level_ud = 95,
  save_rmd = FALSE,
  save_shp = TRUE,
  save_pts = TRUE,
  output_dir = "temp",
  out_format = "pdf_document",
  ...
)
}
\arguments{
\item{params}{List of 19 parameter values used to generate report.}

\item{iterations}{Numeric. Number of iterations to perform for track2KBA::repAssess. More iterations will result in a more accurate assessment, but take longer to run. Default 5.}

\item{level_ud}{Numeric, from 0 to 100. Utilization distribution volume to extract from kernel densities. Defaults to 95.}

\item{save_rmd}{Logical (T/F). Should the .Rmd file used to generate the PDF report be saved as well?}

\item{save_shp}{Logical (T/F). Should the final OPP key area polygon be saved as a shapefile?}

\item{save_pts}{Logical (T/F). Should the raw downloaded Movebank points be saved as a shapefile?}

\item{output_dir}{Output directory for generated files, relative to your working directory. Defaults to 'temp'. If the directory does not exist, the script will create the directory.}

\item{out_format}{Output file format for the knitted document. Defaults to 'pdf_document'.}
}
\description{
This function will generate a KBA report using the 'opp-kba-report' RMarkdown template in \code{OPPtools}, with options to save the generated report as an .Rmd file in addition to the output PDF.
}
\examples{

for (i in 1:nrow(cpf_report_params)) {
  opp_render_report(params = cpf_report_params[i,],
                    kernel_smoother = cpf_report_params$smoother[i],
                    iterations = 10,
                    level_ud = 95,
                    save_rmd = FALSE,
                    save_pts = TRUE,
                    output_dir = 'temp',
                    out_format = 'pdf_document')
                    }

}
