% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPReporting.R
\name{opp_reports_cpf}
\alias{opp_reports_cpf}
\title{Generate reports for central place foraging species tracked in OPP}
\usage{
opp_reports_cpf(
  analysis_table = cpf_report_params,
  login = NULL,
  output_dir = "temp"
)
}
\arguments{
\item{analysis_table}{Data frame that provides all values needed for the RMarkdown template, see details.
Defaults to the cpf_report_params data object that is contained within OPPTools.}

\item{login}{Stored Movebank login credentials if provided, otherwise function
will prompt users to enter credentials.}

\item{output_dir}{Name of of the directory where reports should be saved.}
}
\description{
This function passes a table of parameters to an RMarkdown template to generate summary
reports on tracking data analysis.
}
\details{
Still need to write this up, once function is complete.

The example below will only work if the user has permissions for all the projects listed in cpf_report_params
}
\examples{

print(cpf_report_params)
opp_reports_cpf()

}
