% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{opp_retrieve_mb_cred}
\alias{opp_retrieve_mb_cred}
\title{Retrieve stored Movebank credentials}
\usage{
opp_retrieve_mb_cred()
}
\value{
Returns an object of class "MovebankLogin" which can be passed to
functions in the move package when dowloading data from Movebank.
}
\description{
Retrieve stored Movebank credentials
}
\details{
Looks for OPP-Movebank credentials on keyring. If no credentials are
are found, user is prompted to enter credentials.
}
\examples{

mb_login <- opp_retrieve_mb_cred()
mb_login
}
