% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_sites}
\alias{opp_sites}
\title{Delineate sites of importance for OPP emergency response}
\usage{
opp_sites(
  kernels,
  repr,
  population = NA,
  level_ud = 95,
  thresh = c(10, 50, 75, 95),
  metadata = list(),
  save_shp = FALSE,
  out_name = NA,
  out_dir = NA
)
}
\arguments{
\item{kernels}{Output of class estUDm from either opp_kernels, opp_bbmm, or adehabitatHR::kernelUD.}

\item{repr}{Numeric (0-100). Representativeness of the sample, derived from track2KBA::repAssess.}

\item{population}{Numeric (optional). Total number of individuals at a given study site. Used to calculate the number of individuals that use a given polygon area.}

\item{level_ud}{Numeric (0-100). Utilization distribution level to extract from kernels prior to calculating high-use areas. Defaults to 95.}

\item{thresh}{List of percentile threshold values to calculate high use areas. By default, the function will calculate polygons for the 10th, 50th, 75th, and 95th percentile use areas (\code{c(10, 50, 75, 95)})}

\item{metadata}{List of additional metadata you wish to append to the output. E.g., \code{c("smoother" = "href", "smoother_km" = 3.5)}}

\item{save_shp}{Logical (T/F). Should the output be saved as a shapefile (.shp)?}

\item{out_name}{Character string. Filename for output shapefile. Defaults to 'opp_sites' if none provided and save_shp = TRUE.}

\item{out_dir}{Character string. Output directory for saved shapefile, relative to the current working directory. Defaults to the current working directory if no value provided.}
}
\value{

}
\description{
This function generates key biological areas that are usable in the context of OPP
risk mitigation. By default, the function will delineate areas that are used by at least 10\%, 50\%,
75\%, and if possible, 95\% of the population. If a population number is provided, it will also
calculate an estimate of the number of individuals using a given area. By providing a list of
percentile thresholds, the user can also specify their own custom percentile threshold values.
}
