% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_step}
\alias{opp_step}
\title{Calculate median step length for kernel density estimates}
\usage{
opp_step(data)
}
\arguments{
\item{data}{SpatialPointsDataFrame containing projected tracking data,
with an ID field indicating unique trips}
}
\description{
Calculate median step length for kernel density estimates
}
