% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_study_summary}
\alias{opp_study_summary}
\title{Get summary of Movebank studies associated with OPP}
\usage{
opp_study_summary(project_id = opp_mb_projects$mb_project_num)
}
\arguments{
\item{project_id}{List of Movebank project numbers, defaults to projects included in opp_mb_projects}
}
\value{
A data frame with the following fields:
\describe{
\item{Movebank project}{Movebank project number}
\item{Latin name}{Species latin name}
\item{Common name}{Species common name}
\item{Study latitude}{Main study latitude, decimal degrees}
\item{Study longitude}{Main study longitude, decimal degrees}
\item{Sensor type}{Sensors available in Movebank}
\item{Tags}{Number of tags in the study}
\item{Individuals}{Number of individuals in the study}
\item{Deployments}{Number of deployments in the study}
\item{First location}{Date of earliest location}
\item{Last location}{Date of latest location}
}
}
\description{
Get summary of Movebank studies associated with OPP
}
\details{
Users must have a Movebank account and permissions to access specific Movebank projects for this function to work.
If you do now have access to the project, then you can request access through Movebank. If
project data are visible to the publis, you may still need to go to Movebank first to
accept the study license before data are available to your user name.
}
\examples{
opp_study_summary()
}
