% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPPlots.R
\name{plot_interp_dist}
\alias{plot_interp_dist}
\title{Plot interpolated tracks obtained using ctcrw_interpolation()}
\usage{
plot_interp_dist(data, showPlots = T, plotsPerPage = 4)
}
\arguments{
\item{data}{Output from ctcrw_interpolation().}

\item{showPlots}{Logical (T/F), should plots be printed.}

\item{plotsPerPage}{Numeric indicating the number of individuals to include
in a single plot. Defaults to 1.}
}
\value{
A list of ggplot objects. If plotsPerPage == 1, then each object is
named after the corresponding deployment ID in data. If plotsPerPage > 1 then
the function returns an unnnamed list of ggplot objects.
}
\description{
Plots the results of ctcrw_interpolation(), with DateTime on the x-axis
and ColDist. Raw GPS data is plotted in purple and interpolated locations are shown in
teal.
}
