% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPPlots.R
\name{plot_trip_dist}
\alias{plot_trip_dist}
\title{Plot trips identified using opp_get_trips()}
\usage{
plot_trip_dist(
  data,
  plotsPerPage = 1,
  showPlots = T,
  innerBuff = NULL,
  returnBuff = NULL
)
}
\arguments{
\item{data}{SpatialPointsDataFrame object output from opp_get_trips}

\item{plotsPerPage}{Numeric indicating the number of individuals to include
in a single plot. Defaults to 1.}

\item{showPlots}{Logical (T/F), should plots be printed.}

\item{innerBuff}{Minimum distance (km) from the colony to be in a trip.
Used to label trips as 'Non-trip'. Defaults to 5}

\item{returnBuff}{Outer distance (km) to capture trips that start and end
away from the colony. Used to label trips as 'Incomplete'. Defaults to 20.}
}
\value{
A list of ggplot objects. If plotsPerPage == 1, then each object is
named after the corresponding deployment ID in data. If plotsPerPage > 1 then
the function returns an unnamed list of ggplot objects.
}
\description{
Plots the results of opp_get_trips(), with DateTime on the x-axis and ColDist
on the y-axis. Points coloured based on Type
}
