% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPReporting.R
\name{render_diagnostic}
\alias{render_diagnostic}
\title{Render diagnostic report for a given species}
\usage{
render_diagnostic(
  params,
  save_rmd = FALSE,
  output_dir = "temp",
  out_format = "html_document",
  ...
)
}
\arguments{
\item{params}{List of 17 parameter values used to generate report.}

\item{save_rmd}{Logical (T/F). Should the .Rmd file used to generate the PDF report be saved as well?}

\item{output_dir}{Output directory for generated files. Defaults to 'temp'. If the directory does not exist, the script will create the directory.}
}
\description{
This function will generate a diagnostic report using the 'opp-diagnostic-report' RMarkdown template in \code{OPPtools}, with options to save the generated report as an .Rmd file in addition to the output PDF.
}
