% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPReporting.R
\name{render_kba}
\alias{render_kba}
\title{Render KBA report for a given species}
\usage{
render_kba(
  params,
  kernel_smoother = "href/2",
  iterations = 5,
  level_ud = 95,
  save_rmd = FALSE,
  save_shp = TRUE,
  output_dir = "temp",
  out_format = "html_document",
  ...
)
}
\arguments{
\item{params}{List of 16 parameter values used to generate report.}

\item{kernel_smoother}{One of four options for kernel smoothers in kernel calculations: "href", "href/2", "step", or "bbmm". Defaults to href/2.}

\item{iterations}{Numeric. Number of iterations to perform for track2KBA::repAssess. More iterations will result in a more accurate assessment, but take longer to run. Default 5.}

\item{level_ud}{Numeric, from 0 to 100. Utilization distribution volume to extract from kernel densities. Defaults to 95.}

\item{save_rmd}{Logical (T/F). Should the .Rmd file used to generate the PDF report be saved as well?}

\item{output_dir}{Output directory for generated files. Defaults to 'temp'. If the directory does not exist, the script will create the directory.}
}
\description{
This function will generate a KBA report using the 'opp-kba-report' RMarkdown template in \code{OPPtools}, with options to save the generated report as an .Rmd file in addition to the output PDF.
}
