% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{sum_trips}
\alias{sum_trips}
\title{Calculate trip summaries}
\usage{
sum_trips(data)
}
\arguments{
\item{data}{Trip data ouptut from opp_get_trips() or ctcrw_interpolation().}
}
\description{
\code{sum_trips} quickly calculates summary information
such as maximum distance from the colony, trip start time,
trip end time, and trip duration for each individual trip ID.
The function accepts outputs from either \code{opp_get_trips} or
\code{ctcrw_interpolation}. If interpolated data are provided, the
output provides a summary of interpolated trips.
}
\examples{
my_data <- opp_download_data(study = c(1247096889),login = NULL, start_month = NULL,
                            end_month = NULL,season = NULL)

my_track2kba <- opp2KBA(data = my_data)

my_trips <- opp_get_trips(data = my_track2kba, innerBuff  = 5, returnBuff = 20,
                         duration  = 2, gapLimit = 100, missingLocs = 0.2,
                         showPlots = TRUE)

my_interp <- ctcrw_interpolation(data = my_trips,
                                site = my_track2kba$site,
                                type = c('Complete','Incomplete'),
                                timestep = '10 min',
                                showPlots = T,
                                theta = c(8,2)
)

sum_trips(my_trips)
sum_trips(my_interp)

}
