% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{ud_stack}
\alias{ud_stack}
\title{Stack kernel UD objects into a population level kernel}
\usage{
ud_stack(data, weights)
}
\arguments{
\item{data}{List of estUDm objects. Output from either opp_kernel or opp_bbmm.}

\item{weights}{Optional list of weights.}
}
\description{
This function accepts a list of estUDm objects
as output by either opp_kernel or opp_bbmm, then rasterizes
and stacks them to produce a population-level kernel. An optional
list of weights can be supplied to emphasize certain tracks over others.
}
