% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{ud_vol}
\alias{ud_vol}
\title{Extract 50\% and 99\% volume contours from kernels}
\usage{
ud_vol(data, lowerVol = 50, upperVol = 99)
}
\description{
This function allows the user to reclassify utilization distribution kernels
to highlight two specified kernel volumes. By default, this function classifies the 50\%
and 99\% kernel volumes.
}
