% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{bbox_at_zoom}
\alias{bbox_at_zoom}
\title{Defines a bounding box at a specified zoom level around a set of coordinates}
\usage{
bbox_at_zoom(locs, zoom_level = 7)
}
\arguments{
\item{locs}{An sf or sp points object. The bounding box will be centered on the mean value of
these points}

\item{zoom_level}{Numeric. Specifies how zoomed in the bounding box should be, eg. 1 = whole world and 4 = 1/4 of world}
}
\value{
an object with class "bbox" containing four values: xmin, ymin, xmax, and ymax.
Values will be in units of locs (either decimal degrees or meters).
}
\description{
This function returns a bounding box at a user defined zoom level
centered on the mean value of a set of coordinates.
}
\references{
Code adapted from: https://www.r-bloggers.com/2019/04/zooming-in-on-maps-with-sf-and-ggplot2/
}
