% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{createGrid}
\alias{createGrid}
\title{Create a base grid for bbmm kernel calculations}
\usage{
createGrid(data, m = 1000, interpolated = FALSE)
}
\arguments{
\item{data}{Output from either opp_get_trips or ctcrw_interpolation.}

\item{m}{Numeric. Resolution in meters of grid cells.}

\item{interpolated}{Logical (T/F). If output from ctcrw_interpolation is provided, should the raw or interpolated data be used for calculating the grid extent? This parameter is ignored if opp_get_trips data is provided.}
}
\description{
This helper function takes the output from either opp_get_trips
or ctcrw_interpolation and produces a base grid from the trips.
Grid resolution is provided in meters. If ctcrw_interpolation
output is provided, by default the extents are calculated from
the raw (non-interpolated) data.
}
