% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{mb_cols}
\alias{mb_cols}
\title{List additional columns available for download}
\usage{
mb_cols()
}
\value{

}
\description{
This function lists additional column names you can pass to the \code{addl_cols} param in \code{opp_download_data()}. Note that \code{opp_download_data()} will fail if you pass columns that don't exist in the particular study dataset you are downloading.
}
\examples{
 \dontrun{
mb_cols()
}
}
